Add `BuildIl2CppTask` for Il2Cpp Compilation in Unity Integration
